Merged PR 645228: Change to allow upstream protocol to be specified
This change allows user to specify an UpstreamProtocol while using the Edge Runtime.

**Current behavior:**
Currently edge runtime (both Agent and hub) use AMQP over TCP to connect to IoTHub, and fall back to AMQP over WS if that fails. This has the following disadvantages -
- The fallback to WS takes time, so the startup time of the edge runtime increases (customers have actually complained about this)
- The service will soon support MQTT for Modules. But there is no way for a customer to use MQTT (if they want to, for some reason)

**Change:**
This change allows the user to specify the protocol they want to use for communicating with the IoTHub.
- The user specifies the protocol to be used in iotedgectl as the --upstreamProtocol flag
- IotEdgeCtl subsequently adds an environment variable called UpstreamProtocol to EdgeAgent.
- If specified, the EdgeAgent uses the specified protocol to connect to IoTHub. It also injects an environment variable called UpstreamProtocol to all modules it starts.
- The EdgeHub also uses the specified protocol (if any) to connect to IoTHub on behalf of all devices/modules that connect to it.

Currently the EdgeHub and EdgeAgent allow specifying one of Amqp, AmqpWs, Mqtt and MqttWs protocols. But iotedgectl currently only supports Amqp or AmqpWs. It will support Mqtt when the service supports that.

Related work items: #1826760, #1832409