Merged PR 674496: Add Rust docker client crate
This commit includes the implementation of a crate that contains code for interfacing with Docker. This has support for both HTTP endpoints and Unix Domain Sockets via the `hyperlocal` crate. In general the library is extensible using `hyper`'s "connector" mechanism so we should theoretically be able to use this with Windows named pipes as well without any changes (I haven't tested this though). There are some caveats:

1. The code doesn't compile cleanly without warnings. I added a cargo config file to the crate but that doesn't seem to have any effect when you build from the `edgelet` folder. It seems to work only if you build from `edgelet/docker-rs`.
2. I removed code from the `edgelet/docker-rs/src/api` folder that wasn't relevant for Edge (Docker Swarm related code for example). But the model types that live in `edgelet/docker-rs/src/models` haven't been pruned. That is somewhat complicated because we have to do a type dependency analysis to trim it down to exactly what we need. This might be worth doing because it'll cut down on build times.
3. There are some auto-generated struct types with names like `InlineResponse2001` which is less than ideal. Fixing this will require changes to the code generation tool. We can get to that later if needed.
4. The official Docker [swagger spec](https://docs.docker.com/engine/api/v1.36/swagger.yaml) doesn't quite work with the code generation tool. I am using a tweaked version that I picked from [here](https://raw.githubusercontent.com/masaeedu/moby/27919-fixswagger/api/swagger.yaml). The basic issue is that the official spec uses certain inline type definitions that the tool doesn't like. I have checked-in this tweaked version of the file as well.

I had to tweak the swagger code generation template to add support for domain sockets. That lives on GitHub in a fork of the project here - https://github.com/avranju/swagger-codegen/tree/support-rust-uds. I also had to manually tweak the generated code a bit. So if we need to do this again in the future we may need to do some intelligent code-diffing post-generation. The changes are minimal so it shouldn't be too hard.

Since most of this is generated code, there aren't any tests yet.