Merged PR 664387: Update to stable rustfmt
With the release of rust 1.24.0, `cargo fmt` is now available with the stable compiler. This PR updates the build scripts and README to use stable `cargo fmt`. This means we no longer need to run it in a docker container.

More info here: https://blog.rust-lang.org/2018/02/15/Rust-1.24.html

This version of rustfmt changes the default styles to the latest default style. This required rustfmt to run when checking this change so that build will pass.