Merged PR 688225: Update docker-rs model definitions
It turns out that the default behavior for JSON serialization with the `serde_json` crate is that if you have a field in a struct which is an `Option<T>` then it will produce a `null` value in the generated JSON if the field value is `None`. This is not ideal since we get large JSON strings with mostly just `null` values for sparsely populated structures that happen to have lots of optional fields. This PR adds the `skip_serializing_if = "Option::is_none"` attribute for all `Option<T>` fields in all model structs.

This PR also adds an implementation for the `FromStr` trait for `ContainerCreateBody` because we need to be able to de-serialize from a string value when we get "create options" from the agent twin. This will be used from the `docker-mri` crate later.