Merged PR 688256: Add serde utils
This PR adds the `edgelet-utils` crate and adds a serde util function for customizing deserialization of structs that have fields that can be a string value or a JSON map object - like the "createOptions" field in the module JSON for edge agent. The implementation has been adapted from the sample from https://serde.rs/string-or-struct.html.