Merged PR 691217: Added Edgelet proto interface and E2E security document
What started of as a project to come up with a proto interface between Edgelet <--> Modules, it quickly became apparent that there needs to be an investigation of how security is to be implemented E2E. It began with how will the GRPC interfaces be called from modules and how to integrate with the SDK, followed by how do the requests get handled in the Edgelet. So with that said, this PR has a doc and proto interface.

A few details are still TBD and have been called out but more less most of everything needed for GA is covered along with a path towards X.509 auth.

- For the benefit of reviewers, I would suggest clicking the MD file in VSTS followed by the preview button :)