Merged PR 704169: Launch the [Deploy] tests from a bash script
These changes get the [Deploy] tests working in the service deployment build and the release. I was having trouble invoking the tests directly from an inline bash script task in VSTS, so I created a script that can be called from the task instead.

I'm still tempted to ditch the script + XUnit test in favor of a simple app because it's so difficult to pass arguments to an XUnit test (environment variables are the solution), and because we really aren't using any features of the test framework at this point (only one test, no asserts). But I didn't go that far in this change...maybe another time. If we end up adding several more tests, the test framework might end up being the best option because it gives us discovery.

I also added the option for the test to look up container registry credentials in key vault. This is a feature I didn't end up using, but we'll probably use it soon when we make the service deployment build more flexible (able to take a registry argument).

Once this is in, I can save my changes to the build/release definitions.