Merged PR 715703: Amqp support in EdgeHub
**Overview**
This PR contains most of the code for AMQP support in the EdgeHub (details below).
This is a largish PR, but my previous attempts to break it down into smaller chunks resulted in repeated and invasive refactoring, which was slowing down the development process. To help with the PR, I have added comments wherever necessary and also included some notes to help get an idea of the structure of the code. Plus I will be happy to go over the code in person, if anyone wants.

**Details**
- Added support for Twins, direct methods and messages to modules
- Refactored the code to make it easier to understand and to avoid any duplication
- Modified existing tests and added new tests to cover most of the important parts.

**Code structure**
The Amqp protocol head logic is contained in the following important parts -
- Each connection contains several links, one for each operation - Cbs, Events, C2D, module messages, twin sending, twin receiving, method sending and method receiving
- Each link has a link handler to manage its functionality.
- SendingLinkHandler and ReceivingLinkHandler are the base classes that contain the logic for sending and receiving messages from the client.
- Each connection has a connection handler object. It keeps track of the active links, and contains the device listener and identity of the connection.
- There are 3 message converters for converting between IMessage and AmqpMessage for telemetry, twins and methods.

**Work remaining**
- Adding e2e tests, and possibly more unit tests
- Need to handle logic to receive and store out of band tokens from the client on the cbs link.

Related work items: #1907873, #1907887