Merged PR 766201: Move status into runtime state
Moves `status` into `ModuleRuntimeState` to prevent having to make two async calls per module to retrieve all of the required data for the management api.

In order to implement the http api for management, we need all of the status information. The current design would require two async calls, one to get status, the other to get the runtime state. This PR moves status into runtime state to more closely align with the http api and to reduce the number of async calls to one.