Merged PR 769936: Add more functions to HSM crates
Implementing more functionality based on the newer HSM API.

Now for the common HSM functionality, we have an object called Hsm::Crypto, which can be instantiated and the crypto operations can be called.

Added a "free buffer" implementation for TPM, X509 and HsmCrypto interfaces - for ensuring data allocated in one runtime is freed in that runtime.