Merged PR 774232: Package IotEdgeQuickstart as a tarball for linux-arm
Currently the buildBranch.sh script copies the self-contained linux-arm IotEdgeQuickstart directory into target/publish. When  the CD release copies the directory to the Pi it loses all file permissions (e.g. the executable bit on binaries), so the app doesn't run. Plus it's not compressed so the copy operation it takes longer.

These changes create a tarball and copy it to target/publish. I also increased several timeouts inside IotEdgeQuickstart because things take longer on Pi.