Merged PR 784374: Rename docker-mri
The convention in the project seems to be to prefix all edgelet specific crates with the string `edgelet`. Since `docker-mri` was one of the very first crates created, we hadn't established this convention yet. Fixing the anomaly with this commit. If you don't like `edgelet-docker`, let me know and I can rename it to something else.