Merged PR 780025: Added support for SAS token signing in the C based HSM lib
Changes include an implementation of the shared access key based sign and derive and sign operations and the overall framework upon with other HSM API and functionality (ex. certificate APIs) will be delivered.

Code also includes umock based unit tests and the overall C coding style is consistent with Gateway V1.

Note: The crypto store API is currently in memory based and this will be updated to a file based storage in another smaller PR.