Merged PR 780173: EdgeAgent: Add support for IEnvironment for Edgelet
This is a large PR, and slightly complicated. If anyone is interested, I am happy to give an in-person overview of the changes so that you can have full context on what is going on.
Here are the main points at a high level -
- Redesigned support for getting environment information from Edgelet.
- In this design, the deployed configuration is stored locally, and used for the next reconcile to get the environment information.
- Cleaned up IEnvironment interface

Note: Still working on tests (both existing and new - sending this out early so that folks can take a look at it).