Merged PR 816058: Add new request type for hyper servers
Updates the hyper servers to use the request/response types from the `http` crate. This is in preparation for implementing ACL'ing using the pid. These types have an `Extensions` map which will be used to add the pid. The current hyper request/response has no good way of including context.

This PR is large but there should be no functional changes.