Merged PR 827068: Move iotedged main to the lib
Move iotedged main to the lib. This introduces a `Main` struct which takes a `Settings` struct and provides a `run_until` method. The main method then constructs this struct and runs it.

The reason for this change is so that we can eventually add integration tests to the start up of main. This is impossible with the previous setup.

This PR has no functional changes, just moves a bunch of code around.