Merged PR 795602: Add script to initialize an IoT Core device to run Azure IoT Edge
This script is distributed via an aka.ms link to device owners to run as part of the [quickstart guide](https://docs.microsoft.com/en-us/azure/iot-edge/quickstart). It currently doesn't exist anywhere in this repository. This change checks in the live version of the script with a single additional change to rename get-pip's initialization script  from ```pip.py``` to ```dep-bootstrapper.py``` due to an issue with get-pip where the installation fails if the installer script contains the string "pip".