Merged PR 838905: Use generation id when deriving keys for module identities
This is the first in a series of commits that add support for generation ID when generating and using module keys. This commit splits the module identity creation process into 2 steps:

  1. Create  an IoT Hub module with the authentication type set to `None`. This will cause the hub to create a module and return its generation ID.
  2. Update the newly created module with SAS tokens derived from the device key along with the generation ID.

Future commits will add support for generation IDs in the `sign` API and also inject the generation ID as an environment variable in modules.