Merged PR 851889: Upgrade arm32v7 base image to .NET Core 2.1 RC1
We upgraded the .NET Core compile used by the build system to 2.1 RC1 (was 2.1 preview 2), but didn't upgrade the ARM docker base image, so ARM containers weren't working correctly. This change updates the ARM docker images. It also changes the "Mode" variable in Edge Agent from "iotedged" back to "docker"; we can change it back to "iotedged" once that mode is working end-to-end (and it can by modified locally during development).