Merged PR 868463: Add iotedge logs command
Add iotedge logs command:

```
$ ./target/debug/iotedge logs --help
iotedge-logs
Fetch the logs of a module

USAGE:
    iotedge logs [FLAGS] [OPTIONS] <MODULE>

FLAGS:
    -f, --follow     Follow output log
    -h, --help       Prints help information
    -V, --version    Prints version information

OPTIONS:
        --tail <NUM>    Number of lines to show from the end of the log [default: all]

ARGS:
    <MODULE>    Sets the module identity to get logs
```

This is implemented by connecting the logs stream from the docker engine to the management api. The tricky part is parsing this stream because docker puts a header to specify if the logs were written to stdout or stderr. The iotedge command parses this header and writes to the appropriate output.