Merged PR 889166: Set ManagedBy for all Identities created by Edge runtime
For public preview, IoTHub service would set the ManagedBy field for all modules to IotEdge automatically. With the next deployment, the service will stop doing that. So the edge runtime needs to set ManagedBy for all module identities created.

ManagedBy is used to identify which identities should be deleted by the edge runtime if not being used.