Merged PR 895067: Fixing certificate expiration times
Adding changes ensuring certificate expiration never exceed the issuing certificates expiration along with tests.

This was done in the past using OpenSSL ASN1_tim_diff APIs but when trying out against different OpenSSL versions, it was found that this API does not exist in older revisions. The solution implemented here deals with this in a consistent manner across any OpenSSL revision.