Merged PR 898386: Ensure Edge hub runs on container's PID 1
The Edge Hub container currently calls a script, `start.sh`, which starts the application as a non-root-priviledged user via the `su` command. Unfortunately su runs a separate shell, so the app runs in a child process. This leads to two problems:
1. The app won't be authorized to call edgelet APIs like "certificate/server" (get server certificate) because its PID doesn't match the PID docker records. We may want to look at authorizing child PIDs (using `docker top`), but that's a separate issue...
2. The app won't be able to shut down gracefully because it won't receive the signals that are sent to PID 1 when docker stops the container.

We could fix this by using `gosu` (or `su-exec` on Alpine). But since we've been wanting to get rid of our `start.sh` script anyway, I simply changed the Dockerfile to follow the same pattern as our other module Dockerfiles. It uses the script form of `CMD`, but with the `exec` command to ensure the dotnet app runs in PID 1.