Merged PR 896500: Deprovision the device when settings (config + environment) changes
These changes will make iotedged save the settings on each boot. Before saving the settings, iotedged will compare the saved settings with the new settings. If a difference is detected, the device will be deprovisioned i.e. master encryption key will be destroyed and all containers will be removed. After deprovisioning, the device will be reprovisioned with the new settings i.e. new master encryption key is generated and settings are cached again. The settings are stored as a hash and not in clear text. iotedged will treat the very first boot (when no backed up state exists) as a reprovision event.