Merged PR 895699: Caching tokens for offline authentication
Note: This PR was already approved earlier, but was merged into a private branch. https://msazure.visualstudio.com/One/_git/Azure-IoT-Edge-Core/pullrequest/875130?_a=overview
This change is to merge it with master (to make it easier to maintain and release the code). There are minor changes from the PR above.

These are the changes requested by Schlumberger - allowing EdgeHub to cache tokens so that clients that have authenticated before, can re-connect when offline and continue their operations, using the same token.
- Added logic to cache tokens
- Allowed cloud proxy to be created without needing an active IotHub connection

Pending -
- Encrypting cached tokens