Merged PR 910451: EdgeAgent: Report status only if no commands have been skipped
Issue: If a module deployment fails, the EdgeAgent goes into a backoff for module. Then the next time around, if the module commands are skipped, the Agent thinks the deployment was "successful" (if there are no other errors) and reports as such.

Fix: Report success only if no commands were skipped and all commands completed successfully.