Merged PR 930005: Adding a tool to check the Git submodule tree
Adding a tool to check that the Git submodule tree has the same
commit for all referenced remotes.

Working on tests - trying to figure out how to do integration tests
nicely.

This doesn't work on Windows, the combined paths (repo/submodule) we
get for submodules are too long for allowed Windows paths.