Add script to uninstall IoT Edge on Windows (#81)
With this change, IotEdgeSecurityDaemon (renamed from Install-SecurityDaemon.ps1) exports a new function, Uninstall-SecurityDaemon, which takes the following actions:

- Stops and removes the Windows service
- Stops any running containers that are marked with the label "net.azure-devices.edge.owner": "Microsoft.Azure.Devices.Edge.Agent"
- Deletes C:\ProgramData\iotedge and removes it from the system PATH
- Deletes the registry key we set for logging (HKLM:\SYSTEM\CurrentControlSet\Services\EventLog\Application\iotedged)
- Removes the firewall exceptions that were set up for ports 15580 and 15581

If your Windows machine has a partial installation (e.g. the service was stopped/removed and the iotedge directory was deleted but firewall exceptions, registry key, running containers, and system PATH weren't updated), you can use the -Force flag to go through all the steps on a best-effort basis.

Output is minimal, but you can see details if you include the standard -Verbose flag.

This change also fixes a bug in the install function where it doesn't find the Docker service in certain cases. The fix is to recognize both the "Docker" and "com.docker.service" service names.