Expose ability to do plain AMQP over TLS
EdgeHub's current default AMQP transport mode is TCP with TLS upgrade.
This is not convenient for network debugging because conventional tools
cannot decrypt these packets. Exposing an option to do AMQP over TLS
enables someone to decrypt the AMQP traffic between a module and EdgeHub.
This can be done by providing EdgeHub's server certificate to a tool
like Wireshark or tshark.

This is really useful for debugging issues around token refreshes,
packet sizes and other details of the protocol.