Make get_proxy_uri test consistent with expected value.
Considering a https_proxy=https://my-proxy:port the
`get_proxy_uri_recognizes_https_proxy` test will fail because it will
compare:

left: https://my-proxy:port/
right: https://my-proxy:port

This is caused due to the get_proxy_uri parses the `https_proxy` with
`parse::<Uri>()`. The fix is to parse the `proxy_val` in the same way, so
the comparison can be consistent.